Refactor: Lint 및 prettier 적용중
