+++ backend/src/controllers/feedback/element.js
@@ -12,17 +12,17 @@ async function get(req, res, next) {
 
 async function create(req, res, next) {
   const {
-	content, //
-	apply_state,
-	image_id,
-	feedback_id,
+    content, //
+    apply_state,
+    image_id,
+    feedback_id,
   } = req.body;
 
   const element = await ElementService.create(
-	content, //
-	apply_state,
-	image_id,
-	feedback_id,
+    content, //
+    apply_state,
+    image_id,
+    feedback_id,
   );
 
   res.status(200).json(element);
@@ -31,19 +31,18 @@ async function create(req, res, next) {
 async function update(req, res, next) {
   const element_id = req.params.element;
   const {
-    id,
-	content,
-	apply_state,
-	image_id,
-	feedback_id,
+    content, //
+    apply_state,
+    image_id,
+    feedback_id,
   } = req.body;
 
   const element = await ElementService.update(
     element_id, //
-	content,
-	apply_state,
-	image_id,
-	feedback_id,
+    content,
+    apply_state,
+    image_id,
+    feedback_id,
   );
 
   res.status(200).json(element);
