+++ backend/src/routes/user/provider.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import 'express-async-errors';
 
-import ProviderController from '../../controllers/user/provider.js';
+import { ProviderController } from '../../controllers/index.js';
 
 const router = express.Router();
 
@@ -11,7 +11,6 @@ router.get('/:provider', ProviderController.get);
 router.post('/', ProviderController.create);
 router.put('/:provider', ProviderController.update);
 router.delete('/:provider', ProviderController.remove);
-
 
 /* Team (Equal FK) */
